Split client.py into multiple files
