+++ azure-kusto-data/azure/kusto/data/__init__.py
@@ -2,5 +2,7 @@
 # Licensed under the MIT License.
 
 from ._version import VERSION as __version__
-from .client import KustoClient, KustoConnectionStringBuilder, ClientRequestProperties
+from .client import KustoClient
+from .client_request_properties import ClientRequestProperties
+from .kcsb import KustoConnectionStringBuilder
 from .data_format import DataFormat
