+++ azure-kusto-data/azure/kusto/data/client_request_properties.py
@@ -1,0 +1,51 @@
+import json
+from typing import Any
+
+from .kcsb import _assert_value_is_valid
+
+
+class ClientRequestProperties:
+    """This class is a POD used by client making requests to describe specific needs from the service executing the requests.
+    For more information please look at: https://docs.microsoft.com/en-us/azure/kusto/api/netfx/request-properties
+    """
+
+    results_defer_partial_query_failures_option_name = "deferpartialqueryfailures"
+    request_timeout_option_name = "servertimeout"
+    no_request_timeout_option_name = "norequesttimeout"
+
+    def __init__(self):
+        self._options = {}
+        self._parameters = {}
+        self.client_request_id = None
+        self.application = None
+        self.user = None
+
+    def set_parameter(self, name: str, value: str):
+        """Sets a parameter's value"""
+        _assert_value_is_valid(name)
+        self._parameters[name] = value
+
+    def has_parameter(self, name: str) -> bool:
+        """Checks if a parameter is specified."""
+        return name in self._parameters
+
+    def get_parameter(self, name: str, default_value: str) -> str:
+        """Gets a parameter's value."""
+        return self._parameters.get(name, default_value)
+
+    def set_option(self, name: str, value: Any):
+        """Sets an option's value"""
+        _assert_value_is_valid(name)
+        self._options[name] = value
+
+    def has_option(self, name: str) -> bool:
+        """Checks if an option is specified."""
+        return name in self._options
+
+    def get_option(self, name: str, default_value: Any) -> str:
+        """Gets an option's value."""
+        return self._options.get(name, default_value)
+
+    def to_json(self) -> str:
+        """Safe serialization to a JSON string."""
+        return json.dumps({"Options": self._options, "Parameters": self._parameters}, default=str)
